@@ -13,33 +13,33 @@
   "dependencies": {
     "bootstrap": "^4.6.1",
     "bootstrap-vue": "^2.21.2",
-    "firebase": "^9.6.6",
+    "firebase": "^9.6.7",
     "nprogress": "^0.2.0",
-    "vue": "^2.6.14",
+    "vue": "^3.2.31",
     "vue-emoji-picker": "^1.0.3",
-    "vue-router": "^3.5.3",
-    "vuex": "^3.6.2",
+    "vue-router": "^4.0.12",
+    "vuex": "^4.0.2",
     "whatwg-fetch": "^3.6.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.17.5",
     "@babel/eslint-parser": "^7.17.0",
-    "@vue/cli-plugin-babel": "^4.5.15",
-    "@vue/cli-plugin-eslint": "^4.5.15",
-    "@vue/cli-plugin-router": "^4.5.15",
-    "@vue/cli-plugin-vuex": "^4.5.15",
-    "@vue/cli-service": "^4.5.15",
-    "autoprefixer": "^9.8.8",
-    "babel-eslint": "^10.1.0",
-    "cssnano": "^4.1.11",
-    "eslint": "^7.32.0",
+    "@vue/cli-plugin-babel": "^5.0.1",
+    "@vue/cli-plugin-eslint": "^5.0.1",
+    "@vue/cli-plugin-router": "^5.0.1",
+    "@vue/cli-plugin-vuex": "^5.0.1",
+    "@vue/cli-service": "^5.0.1",
+    "autoprefixer": "^10.4.2",
+    "cssnano": "^5.0.17",
+    "eslint": "^8.9.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-vue": "^8.4.1",
     "ftp-deploy": "^2.4.1",
     "prettier": "^2.5.1",
-    "sass": "^1.49.7",
-    "sass-loader": "^10.2.1",
-    "terser-webpack-plugin": "^4.2.3",
+    "sass": "^1.49.8",
+    "sass-loader": "^12.6.0",
+    "terser-webpack-plugin": "^5.3.1",
     "vue-eslint-parser": "^8.2.0",
     "vue-template-compiler": "^2.6.14",
     "webpack-bundle-analyzer": "^4.5.0"
